Split code generation in many code decorators
